fix: move schema fusion util into the model fuser
fuse_input_schema/fuse_model_schema were used by the fuser itself
(model_fuser/fuse.py) but lived under lib/model_manager/schema/,
outside it. Move them into lib/model_manager/utils/torch/model_fuser
as schema_fuse.py and re-export from the fuser's __init__, per
kenns29's review comment ("why is this util in the assembler and not
in the fuser?").
